requirements.txt
black==19.10b0
flake8==3.8.4
isort==5.6.4
opencv-python==4.5.5.62
Pillow==9.0.0
scipy==1.7.3
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
matplotlib==3.5.1
pandas==1.3.5
numpy==1.22.0